# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Studigital API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/captures/suggest-tags": {
      "post": {
        "tags": [
          "captures"
        ],
        "summary": "Suggest Tags",
        "operationId": "suggest_tags_api_v1_captures_suggest_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestTagsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestTagsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/captures/transform": {
      "post": {
        "tags": [
          "captures"
        ],
        "summary": "Transform Capture",
        "operationId": "transform_capture_api_v1_captures_transform_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FlashcardTransformResponse"
                    },
                    {
                      "$ref": "#/components/schemas/NoteTransformResponse"
                    },
                    {
                      "$ref": "#/components/schemas/QuizTransformResponse"
                    }
                  ],
                  "title": "Response Transform Capture Api V1 Captures Transform Post",
                  "discriminator": {
                    "propertyName": "skill_name",
                    "mapping": {
                      "generate_flashcard": "#/components/schemas/FlashcardTransformResponse",
                      "generate_note": "#/components/schemas/NoteTransformResponse",
                      "generate_quiz": "#/components/schemas/QuizTransformResponse"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/captures/": {
      "get": {
        "tags": [
          "captures"
        ],
        "summary": "List Captures",
        "operationId": "list_captures_api_v1_captures__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaptureOut"
                  },
                  "type": "array",
                  "title": "Response List Captures Api V1 Captures  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "captures"
        ],
        "summary": "Create Capture",
        "operationId": "create_capture_api_v1_captures__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/captures/{capture_id}": {
      "get": {
        "tags": [
          "captures"
        ],
        "summary": "Get Capture",
        "operationId": "get_capture_api_v1_captures__capture_id__get",
        "parameters": [
          {
            "name": "capture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_api_v1_artifacts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactOut"
                  },
                  "type": "array",
                  "title": "Response List Artifacts Api V1 Artifacts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "List Tags",
        "operationId": "list_tags_api_v1_tags__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagOut"
                  },
                  "type": "array",
                  "title": "Response List Tags Api V1 Tags  Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/{tag_id}": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "Get Tag",
        "operationId": "get_tag_api_v1_tags__tag_id__get",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserClaims"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_users_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create Profile",
        "operationId": "create_profile_api_v1_users_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "capture_id": {
            "type": "string",
            "title": "Capture Id"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "capture_id",
          "artifact_type",
          "content",
          "created_at"
        ],
        "title": "ArtifactOut"
      },
      "CaptureCreate": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "raw_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Content"
          },
          "media_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Key"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "CaptureCreate"
      },
      "CaptureOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "mode",
          "status",
          "created_at"
        ],
        "title": "CaptureOut"
      },
      "FlashcardPair": {
        "properties": {
          "front": {
            "type": "string",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "title": "Back"
          }
        },
        "type": "object",
        "required": [
          "front",
          "back"
        ],
        "title": "FlashcardPair"
      },
      "FlashcardTransformResponse": {
        "properties": {
          "skill_name": {
            "type": "string",
            "const": "generate_flashcard",
            "title": "Skill Name"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/FlashcardPair"
            },
            "type": "array",
            "title": "Cards"
          },
          "source_summary": {
            "type": "string",
            "title": "Source Summary"
          }
        },
        "type": "object",
        "required": [
          "skill_name",
          "cards",
          "source_summary"
        ],
        "title": "FlashcardTransformResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NoteTransformResponse": {
        "properties": {
          "skill_name": {
            "type": "string",
            "const": "generate_note",
            "title": "Skill Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body_markdown": {
            "type": "string",
            "title": "Body Markdown"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points"
          }
        },
        "type": "object",
        "required": [
          "skill_name",
          "title",
          "body_markdown",
          "key_points"
        ],
        "title": "NoteTransformResponse"
      },
      "QuizQuestion": {
        "properties": {
          "stem": {
            "type": "string",
            "title": "Stem"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 2,
            "title": "Options"
          },
          "correct_index": {
            "type": "integer",
            "title": "Correct Index"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "stem",
          "options",
          "correct_index",
          "explanation"
        ],
        "title": "QuizQuestion"
      },
      "QuizTransformResponse": {
        "properties": {
          "skill_name": {
            "type": "string",
            "const": "generate_quiz",
            "title": "Skill Name"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "skill_name",
          "questions"
        ],
        "title": "QuizTransformResponse"
      },
      "SuggestTagsRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "existing_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Existing Tags"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SuggestTagsRequest"
      },
      "SuggestTagsResponse": {
        "properties": {
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          }
        },
        "type": "object",
        "required": [
          "suggestions"
        ],
        "title": "SuggestTagsResponse"
      },
      "TagOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "mastery_score": {
            "type": "number",
            "title": "Mastery Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color",
          "mastery_score"
        ],
        "title": "TagOut"
      },
      "TransformRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "tier": {
            "type": "string",
            "enum": [
              "free",
              "paid"
            ],
            "title": "Tier",
            "default": "free"
          },
          "skill_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Name"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TransformRequest"
      },
      "UserClaims": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "tier": {
            "type": "string",
            "title": "Tier",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "UserClaims"
      },
      "UserProfileIn": {
        "properties": {
          "learning_type": {
            "type": "string",
            "enum": [
              "language_learner",
              "professional",
              "student",
              "generic"
            ],
            "title": "Learning Type"
          }
        },
        "type": "object",
        "required": [
          "learning_type"
        ],
        "title": "UserProfileIn"
      },
      "UserProfileOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "learning_type": {
            "type": "string",
            "enum": [
              "language_learner",
              "professional",
              "student",
              "generic"
            ],
            "title": "Learning Type"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "learning_type"
        ],
        "title": "UserProfileOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}