# packages/api-client/src/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ArtifactOut = {
    id: string;
    capture_id: string;
    artifact_type: string;
    content: {
        [key: string]: unknown;
    };
    created_at: string;
};

export type CaptureCreate = {
    mode: string;
    raw_content?: (string | null);
    media_key?: (string | null);
};

export type CaptureOut = {
    id: string;
    user_id: string;
    mode: string;
    status: string;
    created_at: string;
};

export type FlashcardPair = {
    front: string;
    back: string;
};

export type FlashcardTransformResponse = {
    skill_name: "generate_flashcard";
    cards: Array<FlashcardPair>;
    source_summary: string;
};

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type NoteTransformResponse = {
    skill_name: "generate_note";
    title: string;
    body_markdown: string;
    key_points: Array<(string)>;
};

export type QuizQuestion = {
    stem: string;
    options: Array<(string)>;
    correct_index: number;
    explanation: string;
};

export type QuizTransformResponse = {
    skill_name: "generate_quiz";
    questions: Array<QuizQuestion>;
};

export type SuggestTagsRequest = {
    text: string;
    existing_tags?: Array<(string)>;
};

export type SuggestTagsResponse = {
    suggestions: Array<(string)>;
};

export type TagOut = {
    id: string;
    name: string;
    color: string;
    mastery_score: number;
};

export type TransformRequest = {
    text: string;
    tier?: 'free' | 'paid';
    skill_name?: (string | null);
};

export type tier = 'free' | 'paid';

export type UserClaims = {
    user_id: string;
    tier?: string;
};

export type UserProfileIn = {
    learning_type: 'language_learner' | 'professional' | 'student' | 'generic';
};

export type learning_type = 'language_learner' | 'professional' | 'student' | 'generic';

export type UserProfileOut = {
    user_id: string;
    learning_type: 'language_learner' | 'professional' | 'student' | 'generic';
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
    input?: unknown;
    ctx?: {
        [key: string]: unknown;
    };
};

export type SuggestTagsApiV1CapturesSuggestTagsPostData = {
    requestBody: SuggestTagsRequest;
};

export type SuggestTagsApiV1CapturesSuggestTagsPostResponse = (SuggestTagsResponse);

export type TransformCaptureApiV1CapturesTransformPostData = {
    requestBody: TransformRequest;
};

export type TransformCaptureApiV1CapturesTransformPostResponse = ((FlashcardTransformResponse | NoteTransformResponse | QuizTransformResponse));

export type ListCapturesApiV1CapturesGetResponse = (Array<CaptureOut>);

export type CreateCaptureApiV1CapturesPostData = {
    requestBody: CaptureCreate;
};

export type CreateCaptureApiV1CapturesPostResponse = (CaptureOut);

export type GetCaptureApiV1CapturesCaptureIdGetData = {
    captureId: string;
};

export type GetCaptureApiV1CapturesCaptureIdGetResponse = (CaptureOut);

export type ListArtifactsApiV1ArtifactsGetResponse = (Array<ArtifactOut>);

export type GetArtifactApiV1ArtifactsArtifactIdGetData = {
    artifactId: string;
};

export type GetArtifactApiV1ArtifactsArtifactIdGetResponse = (ArtifactOut);

export type ListTagsApiV1TagsGetResponse = (Array<TagOut>);

export type GetTagApiV1TagsTagIdGetData = {
    tagId: string;
};

export type GetTagApiV1TagsTagIdGetResponse = (TagOut);

export type GetMeApiV1UsersMeGetResponse = (UserClaims);

export type GetProfileApiV1UsersProfileGetResponse = (UserProfileOut);

export type CreateProfileApiV1UsersProfilePostData = {
    requestBody: UserProfileIn;
};

export type CreateProfileApiV1UsersProfilePostResponse = (UserProfileOut);

export type HealthHealthGetResponse = ({
    [key: string]: (string);
});